Remove old experiments and CTA alerts from Home and Main screens
