+++ DuckDuckGo/HomeControllerDelegate.swift
@@ -29,8 +29,6 @@ protocol HomeControllerDelegate: class {
 
     func homeDidDeactivateOmniBar(home: HomeViewController)
 
-    func showInstructions(_ home: HomeViewController)
-
     func showSettings(_ home: HomeViewController)
     
     func showPrivacyReport(_ home: HomeViewController)
