+++ DuckDuckGo/HomeScreenTips.swift
@@ -52,11 +52,9 @@ class HomeScreenTips {
         self.storage = storage
     }
     
-    func trigger(variantManager: VariantManager = DefaultVariantManager(),
-                 ctaStorage: HomeRowCTAStorage = UserDefaultsHomeRowCTAStorage()) {
+    func trigger() {
         guard storage.isEnabled else { return }
         guard tutorialSettings.hasSeenOnboarding else { return }
-        guard ctaStorage.dismissed else { return }
         guard let tip = Tip(rawValue: storage.nextHomeScreenTip) else { return }
         
         let completion: (Bool) -> Void = { shown in
