+++ DuckDuckGoTests/HomeScreenTipsTests.swift
@@ -28,7 +28,6 @@ class HomeScreenTipsTests: XCTestCase {
     // swiftlint:enable weak_delegate
 
     var storage = MockContextualTipsStorage()
-    let ctaStorage = MockCTAStorage()
     
     var tutorialSettings = MockTutorialSettings()
 
@@ -41,7 +40,7 @@ class HomeScreenTipsTests: XCTestCase {
         let tips = HomeScreenTips(delegate: delegate, tutorialSettings: tutorialSettings, storage: storage)
         XCTAssertNotNil(tips)
 
-        tips?.trigger(ctaStorage: ctaStorage)
+        tips?.trigger()
         XCTAssertEqual(1, delegate.showPrivateSearchTipCounter)
         XCTAssertEqual(0, delegate.showCustomizeTipCounter)
 
@@ -57,24 +56,7 @@ class HomeScreenTipsTests: XCTestCase {
         XCTAssertEqual(0, delegate.showCustomizeTipCounter)
         XCTAssertEqual(0, delegate.showPrivateSearchTipCounter)
 
-        tips?.trigger(ctaStorage: ctaStorage)
-        XCTAssertEqual(0, delegate.showCustomizeTipCounter)
-        XCTAssertEqual(0, delegate.showPrivateSearchTipCounter)
-
-    }
-    
-    func testWhenFeatureEnabledOnboardingShownButCTANotDismissedThenTriggerDoesNothing() {
-
-        storage.isEnabled = true
-        tutorialSettings.hasSeenOnboarding = true
-        let tips = HomeScreenTips(delegate: delegate, tutorialSettings: tutorialSettings, storage: storage)
-        XCTAssertNotNil(tips)
-
-        XCTAssertEqual(0, delegate.showCustomizeTipCounter)
-        XCTAssertEqual(0, delegate.showPrivateSearchTipCounter)
-
-        ctaStorage.dismissed = false
-        tips?.trigger(ctaStorage: ctaStorage)
+        tips?.trigger()
         XCTAssertEqual(0, delegate.showCustomizeTipCounter)
         XCTAssertEqual(0, delegate.showPrivateSearchTipCounter)
 
@@ -99,17 +81,17 @@ class HomeScreenTipsTests: XCTestCase {
         XCTAssertEqual(0, delegate.showPrivateSearchTipCounter)
         XCTAssertEqual(0, storage.nextHomeScreenTip)
 
-        tips?.trigger(ctaStorage: ctaStorage)
+        tips?.trigger()
         XCTAssertEqual(0, delegate.showCustomizeTipCounter)
         XCTAssertEqual(1, delegate.showPrivateSearchTipCounter)
         XCTAssertEqual(1, storage.nextHomeScreenTip)
 
-        tips?.trigger(ctaStorage: ctaStorage)
+        tips?.trigger()
         XCTAssertEqual(1, delegate.showCustomizeTipCounter)
         XCTAssertEqual(1, delegate.showPrivateSearchTipCounter)
         XCTAssertEqual(2, storage.nextHomeScreenTip)
 
-        tips?.trigger(ctaStorage: ctaStorage)
+        tips?.trigger()
         XCTAssertEqual(1, delegate.showCustomizeTipCounter)
         XCTAssertEqual(1, delegate.showPrivateSearchTipCounter)
         XCTAssertEqual(2, storage.nextHomeScreenTip)
@@ -133,10 +115,6 @@ class MockHomeScreenTipsDelegate: NSObject, HomeScreenTipsDelegate {
         didShow(shown)
     }
     
-}
-
-class MockCTAStorage: HomeRowCTAStorage {
-    var dismissed = true
 }
 
 class MockTutorialSettings: TutorialSettings {
