--- DuckDuckGoTests/UserDefaultsHomeRowCTAStorageTests.swift
@@ -1,46 +1,0 @@
-//
-//  UserDefaultsHomeRowCTAStorageTests.swift
-//  DuckDuckGo
-//
-//  Copyright © 2018 DuckDuckGo. All rights reserved.
-//
-//  Licensed under the Apache License, Version 2.0 (the "License");
-//  you may not use this file except in compliance with the License.
-//  You may obtain a copy of the License at
-//
-//  http://www.apache.org/licenses/LICENSE-2.0
-//
-//  Unless required by applicable law or agreed to in writing, software
-//  distributed under the License is distributed on an "AS IS" BASIS,
-//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-//  See the License for the specific language governing permissions and
-//  limitations under the License.
-//
-
-import XCTest
-@testable import Core
-@testable import DuckDuckGo
-
-class UserDefaultsHomeRowCTAStorageTests: XCTestCase {
-
-    let defaultsName = "UserDefaultsHomeRowOnboardingFeatureStorageTests"
-
-    override func setUp() {
-        UserDefaults(suiteName: defaultsName)?.removePersistentDomain(forName: defaultsName)
-    }
-
-    func testWhenDismissedSetThenItIsStored() {
-        createStorage().dismissed = true
-        XCTAssertTrue(createStorage().dismissed)
-    }
-
-    func testWhenNewThenValuesAreDefault() {
-        XCTAssertFalse(createStorage().dismissed)
-    }
-
-    func createStorage() -> UserDefaultsHomeRowCTAStorage {
-        let userDefaults = UserDefaults(suiteName: defaultsName)!
-        return UserDefaultsHomeRowCTAStorage(userDefaults: userDefaults)
-    }
-
-}
